Fix Wrong check in HttpFileTransferSession.terminateSession
HttpFileTransferSession does not contain SipDialogPath. Hence
removed the checks involving dialog path for HttpFileTransferSession.
